+++ src/lib_proxy/light_consensus.mli
@@ -23,6 +23,8 @@
 (*                                                                           *)
 (*****************************************************************************)
 
+module Proof = Tezos_context_sigs.Context.Proof_types
+
 (** A container of input data needed to process a consensus. *)
 type input = {
   printer : Tezos_client_base.Client_context.printer;
@@ -33,13 +35,9 @@ type input = {
       (** The block considered *)
   key : string list;
       (** The key of the context for which data is being requested *)
-  mtree : Tezos_context_sigs.Context.Proof_types.merkle_tree;
-      (** The tree received from the endpoint providing data.
-          It is much smaller than [tree]. *)
-  tree : Local_context.tree;
-      (** The current data tree. The call to [M.consensus]
-      will check that validating endpoints send data
-      which agree with this tree. *)
+  mproof : Proof.tree Proof.t;
+      (** The Merkle proof received from the endpoint providing data.
+          It is much smaller than the whole context. *)
 }
 
 (** [min_agreeing_endpoints min_agreement nb_endpoints] returns