+++ src/lib_shell_services/test_helpers/shell_services_test_helpers.mli
@@ -30,12 +30,9 @@ val raw_context_gen : raw_context QCheck2.Gen.t
 
 val print_raw_context : raw_context -> string
 
-(** [merkle_node_gen] is a {!QCheck2.Gen.t} for [merkle_node]. *)
-val merkle_node_gen : merkle_node QCheck2.Gen.t
+(** [merkle_proof_gen] is a {!QCheck2.Gen.t} for [Proof.tree Proof.t]. *)
+val merkle_proof_gen :
+  (tree t * Tezos_context_memory.Context.tree * string list) QCheck2.Gen.t
 
-val print_merkle_node : merkle_node -> string
-
-(** [merkle_tree_gen] is a {!QCheck2.Gen.t} for [merkle_tree]. *)
-val merkle_tree_gen : merkle_tree QCheck2.Gen.t
-
-val print_merkle_tree : merkle_tree -> string
+val print_merkle_proof :
+  tree t * Tezos_context_memory.Context.tree * string list -> string